requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-tailwind==0.5.0
cssselect2==0.6.0
Django==4.0.4
django-ckeditor==6.4.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-jsonfield==1.4.1
django-phonenumber-field==6.1.0
django-rosetta==0.9.8
fonttools==4.33.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
phonenumbers==8.12.47
phonenumberslite==8.12.46
Pillow==9.1.0
polib==1.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.0.0
zopfli==0.2.1
